Move EmbeddingBacklog related classes to backlog package
