+++ src/main/java/de/uol/pgdoener/civicsage/bootstrap/BootstrapService.java
@@ -2,7 +2,7 @@ package de.uol.pgdoener.civicsage.bootstrap;
 
 import de.uol.pgdoener.civicsage.business.dto.IndexFilesRequestInnerDto;
 import de.uol.pgdoener.civicsage.business.dto.IndexWebsiteRequestDto;
-import de.uol.pgdoener.civicsage.business.embedding.EmbeddingPriority;
+import de.uol.pgdoener.civicsage.business.embedding.backlog.EmbeddingPriority;
 import de.uol.pgdoener.civicsage.business.index.IndexService;
 import de.uol.pgdoener.civicsage.business.index.document.MetadataKeys;
 import de.uol.pgdoener.civicsage.business.source.FileHashingService;
