+++ src/main/java/de/uol/pgdoener/civicsage/business/embedding/backlog/EmbeddingPriority.java
@@ -1,4 +1,4 @@
-package de.uol.pgdoener.civicsage.business.embedding;
+package de.uol.pgdoener.civicsage.business.embedding.backlog;
 
 public enum EmbeddingPriority implements Comparable<EmbeddingPriority> {
     HIGH,
